ajouter statistique + validation de smethodes implementation

+++ src/main/java/com/civil/project/service/MarginaleFrJugeNaissanceServiceImpl.java
@@ -3,7 +3,9 @@ package com.civil.project.service;
 import com.civil.project.dao.JugeNaissanceMargFrRep;
 import com.civil.project.entity.MargJugeNaissFr;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -31,13 +33,20 @@ public class MarginaleFrJugeNaissanceServiceImpl implements MargianleFrJugeNaiss
     //supprimer marginal francais de juge de naissance
     @Override
     public void SupprimerMargFR(int idMarginal) {
-        jugeNaissanceMargFrRep.deleteById(idMarginal);
+        if(MargFrById(idMarginal)!=null){
+            jugeNaissanceMargFrRep.deleteById(idMarginal);
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
+                String.format("il n'y a aucun marginal de juge de naissance avec l'id %d  ",idMarginal));
     }
 
     //chercher le marginal francais de juge de naissance a partir de ID acte de juge de naissance
     @Override
     public List<MargJugeNaissFr> MargFrByIdActeJugeNaissance(int idActeJugeNaissa) {
-
+        if(jugeNaissanceMargFrRep.findByIdActeJugeNais(idActeJugeNaissa).isEmpty())
+        {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,String.format("il n'y a aucun marginal lie avec l'id %d d'acte de juge de naissance ",idActeJugeNaissa));
+        }
         return jugeNaissanceMargFrRep.findByIdActeJugeNais(idActeJugeNaissa);
     }
 
@@ -53,7 +62,7 @@ public class MarginaleFrJugeNaissanceServiceImpl implements MargianleFrJugeNaiss
             margJugeNaissFr=resultat.get();
         }
         else {
-            throw new RuntimeException("Registre non trouve");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,String.format("il n'y a aucun marginal  avec l'id %d  ",id));
         }
         return margJugeNaissFr;
     }
